test(core): split runtime service responsibilities
Split the runtime service test suite into cohesive process, route, supervisor, engine, assessment, and shared-support owners.

Preserve all definitions and tests with AST-identical bodies and original collection order.

Refresh governed runtime and portable documentation projections and renew exact inherited complexity scopes.

(cherry picked from commit 439e90ab3a18a5f3353ff2b27f0b1e6efb283cee)
